# data-api/open-api-specification.yml
openapi: '3.0.0'
info:
  title: 'laa-ccms-data-api'
  version: '1.0.0'
paths:
  /users/{login-id}:
    get:
      tags:
        - users
      summary: 'Get User'
      operationId: 'getUser'
      parameters:
        - name: 'login-id'
          in: 'path'
          required: true
          schema:
            type: 'string'
            example: 'abc123'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userDetail"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /providers/{provider-id}:
    get:
      tags:
        - providers
      summary: 'Get Provider'
      operationId: 'getProvider'
      x-spring-paginated: false
      parameters:
        - name: 'provider-id'
          in: 'path'
          required: true
          schema:
            type: 'integer'
            example: 12345
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/providerDetail"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /lookup/amendment-types:
    get:
      tags:
        - lookup
      summary: 'Get Amendment Type Lookup Values'
      operationId: 'getAmendmentTypeLookupValues'
      x-spring-paginated: true
      parameters:
        - name: 'application-type'
          in: 'query'
          schema:
            type: 'string'
            example: 'ABC'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/amendmentTypeLookupDetail"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /lookup/countries:
    get:
      tags:
        - lookup
      summary: 'Get Countries Lookup Values'
      operationId: 'getCountriesLookupValues'
      x-spring-paginated: true
      parameters:
        - name: 'code'
          in: 'query'
          schema:
            type: 'string'
            example: 'ABC'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/commonLookupDetail"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /lookup/case-status:
    get:
      tags:
        - lookup
      summary: 'Get Case Status Lookup Values'
      operationId: 'getCaseStatusLookupValues'
      x-spring-paginated: true
      parameters:
        - name: 'code'
          in: 'query'
          schema:
            type: 'string'
            example: 'ABC'
        - name: 'copy-allowed'
          in: 'query'
          schema:
            type: 'boolean'
            example: true
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/caseStatusLookupDetail"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
  /lookup/common:
    get:
      tags:
        - lookup
      summary: 'Get a list of common lookup values'
      operationId: 'getCommonLookupValues'
      x-spring-paginated: true
      parameters:
        - name: 'type'
          in: 'query'
          schema:
            type: 'string'
            example: 'CONTACT_TITLE'
        - name: 'code'
          in: 'query'
          schema:
            type: 'string'
            example: 'D'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/commonLookupDetail"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
components:
  schemas:
    baseOffice:
      type: 'object'
      properties:
        id:
          type: 'integer'
        name:
          type: 'string'
    officeDetail:
      allOf:
        - $ref: "#/components/schemas/baseOffice"
      type: 'object'
      properties:
        fee_earners:
          type: 'array'
          items:
            $ref: '#/components/schemas/contactDetail'
    baseProvider:
      type: 'object'
      properties:
        id:
          type: 'integer'
        name:
          type: 'string'
        offices:
          type: 'array'
          items:
            $ref: "#/components/schemas/baseOffice"
    providerDetail:
      allOf:
        - $ref: "#/components/schemas/baseProvider"
      type: 'object'
      properties:
        offices:
          type: 'array'
          items:
            $ref: "#/components/schemas/officeDetail"
    userDetail:
      type: 'object'
      properties:
        user_id:
          type: 'integer'
        login_id:
          type: 'string'
        username:
          type: 'string'
        user_type:
          type: 'string'
        provider:
          $ref: "#/components/schemas/baseProvider"
        firms:
          type: 'array'
          items:
            $ref: "#/components/schemas/baseProvider"
        functions:
          type: 'array'
          items:
            type: 'string'
    contactDetail:
      type: 'object'
      properties:
        id:
          type: 'integer'
          example: 12345
        name:
          type: 'string'
          example: 'a name'
    amendmentTypeLookupDetail:
      allOf:
        - $ref: "#/components/schemas/page"
      type: 'object'
      properties:
        content:
          type: 'array'
          items:
            $ref: '#/components/schemas/amendmentTypeLookupValueDetail'
    amendmentTypeLookupValueDetail:
      type: 'object'
      properties:
        application_type_code:
          type: 'string'
          example: 'ABC'
        application_type_description:
          type: 'string'
          example: 'description'
        cost_limit_cap:
          type: 'string'
          example: '1350.00'
        devolved_powers_indicator:
          type: 'string'
          example: 'Y'
        default_lar_scope_flag:
          type: 'string'
          example: 'Y'
    caseStatusLookupDetail:
      allOf:
        - $ref: "#/components/schemas/page"
      type: 'object'
      properties:
        content:
          type: 'array'
          items:
            $ref: '#/components/schemas/caseStatusLookupValueDetail'
    caseStatusLookupValueDetail:
      type: 'object'
      properties:
        code:
          type: 'string'
          example: 'ABC'
        description:
          type: 'string'
          example: 'a description'
        copyAllowed:
          type: 'boolean'
          example: true
    commonLookupValueDetail:
      type: 'object'
      properties:
        type:
          type: 'string'
        code:
          type: 'string'
        description:
          type: 'string'
        start_date_active:
          type: 'string'
        attribute11:
          type: 'string'
        attribute12:
          type: 'string'
        enabled:
          type: 'boolean'
        default_code:
          type: 'string'
    commonLookupDetail:
      allOf:
        - $ref: "#/components/schemas/page"
      type: 'object'
      properties:
        content:
          type: 'array'
          items:
            $ref: "#/components/schemas/commonLookupValueDetail"
    page:
      type: 'object'
      properties:
        total_pages:
          type: 'integer'
        total_elements:
          type: 'integer'
        number:
          type: 'integer'
        size:
          type: 'integer'